% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activeC.R
\name{as.matrix.const_C}
\alias{as.matrix.const_C}
\title{Extract Matrix}
\usage{
\method{as.matrix}{const_C}(x, ...)
}
\arguments{
\item{x}{A const_C object with field 'mat'.}

\item{...}{Additional parameters. Not used.}
}
\value{
The mat entry of C, a matrix.
}
\description{
Given a const_C object, extracts the actual matrix itself.
}
