% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{C_GP_cpp}
\alias{C_GP_cpp}
\title{Equivalent of \code{\link[activegp]{C_GP}} using RcppArmadillo}
\usage{
C_GP_cpp(design, response, theta, Ki, ct)
}
\arguments{
\item{design}{A matrix of design points, one in each row}

\item{response}{A vector of observations at each design point.}

\item{theta}{vector of lengthscales}

\item{Ki}{inverse covariance matrix}

\item{ct}{Covariance type, 1 means Gaussian, 2 means Matern 3/2, 3 means Matern 5/2}
}
\value{
The active subspace matrix C.
}
\description{
Equivalent of \code{\link[activegp]{C_GP}} using RcppArmadillo
}
\keyword{internal}
