% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hetGP_funcs.R
\name{d1}
\alias{d1}
\title{Get Integerodifferential Quantities}
\usage{
d1(X, x, sigma, type)
}
\arguments{
\item{X}{The design matrix}

\item{x}{The point at which to evaluate.}

\item{The}{error variance of the model.}

\item{A}{string, one of "Gaussian", "Matern5_2", and "Matern3_2" indicating the covariance kernel to use.}
}
\value{
The scalar value given the integrated derivative.
}
\description{
Returns an element related to an integral of a derivative of the desired kernel. See references for details.
}
\references{
Mickael Binois, Robert B. Gramacy, Michael Ludkovski (2017), Practical Heteroscedastic Gaussian Process Modeling for Large Simulation Experiments, Journal of Computational and Graphical Statistics
}
\keyword{internal}
