% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{domain_to_unit}
\alias{domain_to_unit}
\title{Change a function's inputs to live in [-1, 1]}
\usage{
domain_to_unit(f, domain)
}
\arguments{
\item{f}{The function to wrap, should have a single vector-valued input.}

\item{domain}{A list of real tuples, indicating the original domain of the function.}
}
\value{
A function wrapping f.
}
\description{
Given an m dimensional function whose inputs live in bounded intervals [a1, b1], ..., [am, bm], return a wrapped version of the function whose inputs live in [-1, 1], ..., [-1, 1].
}
