% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperparamUQ.R
\name{logLikHessian}
\alias{logLikHessian}
\title{Hessian of the log-likelihood with respect to lengthscales hyperparameters
Works for homGP and hetGP models from the hetGP package for now.}
\usage{
logLikHessian(model)
}
\arguments{
\item{model}{homGP model}
}
\value{
A matrix giving the Hessian of the GP loglikelihood.
}
\description{
Hessian of the log-likelihood with respect to lengthscales hyperparameters
Works for homGP and hetGP models from the hetGP package for now.
}
