% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activeC.R
\name{Lt_GP}
\alias{Lt_GP}
\title{Active Subspace Prewarping}
\usage{
Lt_GP(..., C)
}
\arguments{
\item{...}{Parameters to be passed to C_GP, if C was not provided.}

\item{C}{the result of a call to C_GP. If provided, all other arguments are ignored.}
}
\value{
The matrix Lt which can be used for sensitivity prewarping, i.e. by computing Xw = X %*% Lt.
}
\description{
Computes a matrix square root of C = Lt %*% t(Lt).
}
