% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{n11_2_01}
\alias{n11_2_01}
\title{f:[-1, 1] -> R Becomes f:[0,1] -> R}
\usage{
n11_2_01(f)
}
\arguments{
\item{f}{initial function}
}
\value{
The same function with domain shifted.
}
\description{
f:[-1, 1] -> R Becomes f:[0,1] -> R
}
