% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activeC.R
\name{plot.const_C}
\alias{plot.const_C}
\title{Plot const_C objectc}
\usage{
\method{plot}{const_C}(x, output = c("all", "matrix", "logvals",
  "projfn"), ...)
}
\arguments{
\item{x}{A const_C object, the result of a call to C_GP}

\item{output}{one of \code{"image"} (image of the C matrix), \code{"logvals"} (log-eigen values), 
\code{"projfn"} projected function on first eigen vector or all plots at once (default).}

\item{...}{Additional parameters. Not used.}
}
\description{
Plot const_C objectc
}
