% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activeC.R
\name{print.const_C}
\alias{print.const_C}
\title{Print const_C objects}
\usage{
\method{print}{const_C}(x, ...)
}
\arguments{
\item{x}{A const_C object, the result of a call to C_GP}

\item{...}{Additional parameters. Not used.}
}
\description{
Print const_C objects
}
