% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activeC.R
\name{update_C2}
\alias{update_C2}
\title{Update Constantine's C, using update formula}
\usage{
update_C2(C, xnew, ynew)
}
\arguments{
\item{C}{A const_C object, the result of a call to \code{\link[activegp]{C_GP}}.}

\item{xnew}{The new design point}

\item{ynew}{The new response}
}
\value{
Updated C matrix, a const_C object.
}
\description{
Update Constantine's C, using update formula
}
\references{
N. Wycoff, M. Binois, S. Wild (2019+), Sequential Learning of Active Subspaces, preprint.\cr
}
