% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{W_kappa_ij}
\alias{W_kappa_ij}
\title{Covariance of kernel computations}
\usage{
W_kappa_ij(design, theta, i1, i2, ct)
}
\arguments{
\item{design}{matrix of design points}

\item{theta}{lengthscales}

\item{i1, i2}{index of the derivatives (WARNING: starts at 0)}

\item{ct}{Covariance type, 1 means Gaussian, 2 means Matern 3/2, 3 means Matern 5/2}
}
\value{
The matrix representing the result of the integration.
}
\description{
Computes Int(kappa_i(X, design) . kappa_j(design, X)). This function is preferred for initialization
}
\keyword{internal}
