% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{W_kappa_lk}
\alias{W_kappa_lk}
\title{Covariance of kernel computations}
\usage{
W_kappa_lk(design, theta, i1, i2, ct)
}
\arguments{
\item{design}{matrix of design points}

\item{theta}{lengthscales}

\item{i1, i2}{index of the observations (WARNING: starts at 0)}

\item{ct}{Covariance type, 1 means Gaussian, 2 means Matern 3/2, 3 means Matern 5/2}
}
\value{
The matrix representing the result of the integration.
}
\description{
Creates a submatrix of the same tensor as W_kappa_ij, but this time l,k give the indices of the observations, not variables.
}
\keyword{internal}
