% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{C_GP_cpp}
\alias{C_GP_cpp}
\title{Covariance of kernel computations}
\usage{
C_GP_cpp(design, response, theta, Ki, ct)
}
\arguments{
\item{design}{matrix of design points}

\item{theta}{lengthscales}

\item{Ki}{The inverse covariance matrix}

\item{ct}{Covariance type, 1 means Gaussian, 2 means Matern 3/2, 3 means Matern 5/2}

\item{Kir}{The inverse covariance matrix times the response.}
}
\value{
The matrix representing the result of the integration.
}
\description{
Computes Int(kappa_i(X, design) . kappa_j(design, X)). This function is preferred for initialization
}
\keyword{internal}
